requirements.txt
Werkzeug~=3.1.3
requests~=2.32.3
python-dotenv~=1.0.1
numpy~=2.0.2
matplotlib~=3.10.0
pillow~=11.1.0
ollama~=0.4.7
deepface~=0.0.93
opencv-python~=4.11.0.86
scikit-image~=0.25.1
Flask~=3.1.0